requirements.txt
pandas>=2.2
numpy>=1.26
matplotlib>=3.8
plotly>=5.22
scikit-learn>=1.4
statsmodels>=0.14
pygam>=0.9
requests>=2.32
beautifulsoup4>=4.12
lxml>=5.3
python-dotenv>=1.0
pyyaml>=6.0
jupyter>=1.0
notebook>=7.0
